Refactor into package, fix period parsing, update GUI
