+++ README.md
@@ -1,27 +1,33 @@
-# timecounter
+# WorkTimeTracker
 
-Simple command line utility that sums work time intervals from a log file.
+Utility and GUI for summing work time intervals from a plain text log.
 
-## Usage
+## Installation
 
+Install the dependencies using pip:
+
+```bash
+pip install -r requirements.txt
 ```
+
+## Command line usage
+
+```bash
 python timecounter.py path/to/log.txt
 ```
 
-### Launching the GUI
+This prints the total hours since the last `ОПЛАЧЕНО` entry and, when the log
+contains a following `СЧЕТ ВЫСТАВЛЕН` line, also reports already invoiced hours.
+After displaying the statistics the tool allows you to mark the invoice as paid
+or add new log lines.
 
-```
+## GUI
+
+Launch the graphical interface with:
+
+```bash
 python run_gui.py
 ```
 
-The tool prints the total hours since the last `ОПЛАЧЕНО` entry and, if the
-log contains a subsequent `СЧЕТ ВЫСТАВЛЕН` line, also reports the hours that
-were already invoiced but not yet paid.
-
-After displaying the calculated time, the program interactively offers to:
-
-* mark the invoiced period as paid;
-* mark the last unpaid period as invoiced;
-* append additional log entries.
-
-Any confirmed changes are immediately written back to the log file.
+The GUI displays periods separated by statuses and visual breaks between work
+blocks with dashed lines.
